protect rdtscp. now can run on arm and pentium2
